requirements.txt
aiogram==3.10.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp==3.9.5